feat: implemented payout

+++ src/main/java/com/arthur/blackjack/models/player/Player.java
@@ -5,9 +5,9 @@ import com.arthur.blackjack.models.hand.PlayerHand;
 import java.util.List;
 
 public interface Player {
-    int getBankroll();
-    void subtractFromBankroll(int amount);
-    void addToBankroll(int amount);
+    double getBankroll();
+    void subtractFromBankroll(double amount);
+    void addToBankroll(double amount);
     List<PlayerHand> getHands();
     void addHand(PlayerHand hand);
     void clearHands();
